selling period module

+++ resources/views/edbr/profile.blade.php
@@ -767,7 +767,7 @@ if ($currentMonth >= 4) {
                                                         <div class="title">Leave Balance Days</div>
                                                         @php $cal_leave = (isset($annualleavedetails) && $annualleavedetails['totalLeaveDays']>0 )?$annualleavedetails['totalLeaveDays']:0; 
                                                             $used_leave = $user->used_leave ?? 0;
-                                                            $bal_leave = $cal_leave - $used_leave;@endphp
+                                                            $bal_leave = (int) ($cal_leave - $used_leave);@endphp
                                                         <div class="text">{{$bal_leave ?? 0}}</div>
                                                     </li>
                                                     <li>
@@ -953,7 +953,7 @@ if ($currentMonth >= 4) {
                                                         <div class="title">Balance Days</div>
                                                         @php $cal_leave = (isset($annualleavedetails) && $annualleavedetails['totalLeaveDays']>0 )?$annualleavedetails['totalLeaveDays']:0; 
                                                             $used_leave = $user->used_leave ?? 0;
-                                                            $bal_leave = $cal_leave - $used_leave;@endphp
+                                                            $bal_leave = (int)($cal_leave - $used_leave);@endphp
                                                         <div class="text">{{$bal_leave ?? 0}}</div>
                                                     </li>
                                                     <li>
